refactor: split x86-specific code into x86_ prefixed files in elfanalyzer
- Split syscall_decoder.go: keep generic interfaces (DecodedInstruction,
  MachineCodeDecoder) in place; move X86Decoder implementation to x86_decoder.go
- Rename syscall_decoder_test.go -> x86_decoder_test.go
- Split syscall_numbers.go: keep generic interfaces (SyscallNumberTable,
  SyscallDefinition) in place; move X86_64SyscallTable to x86_syscall_numbers.go
- Rename syscall_numbers_test.go -> x86_syscall_numbers_test.go
- Add file-level nolint:dupl directives to x86_syscall_numbers.go and
  arm64_syscall_numbers.go to suppress structurally identical table warnings
