+++ internal/runner/security/elfanalyzer/x86_syscall_numbers_test.go
